Refactor TensorRecord
