+++ python/python_direct/ir.cpp
@@ -122,116 +122,6 @@ IterDomain
 )");
 }
 
-// Determines which dimensions of a tensor are expanded.
-//
-// A dimension is considered expanded if:
-// 1. It is marked as a broadcast dimension (contiguity is None)
-// 2. It has a non-broadcast size (shape != 1)
-//
-// Parameters
-// ----------
-// shape : vector<int64_t>
-//     The shape of the tensor
-// contiguity : vector<optional<bool>>
-//     The contiguity flags for each dimension. None indicates a broadcast
-//     dimension.
-// stride_order : vector<int64_t>
-//     The stride order of dimensions, mapping logical domain to allocation
-//     domain.
-//
-// Returns
-// -------
-// vector<bool>
-//     A vector of boolean flags indicating which dimensions are expanded.
-//     The order of this vector corresponds to the logical domain of TensorView.
-//
-// Notes
-// -----
-// The function handles the mapping between logical domain and allocation domain
-// using the stride order. For each dimension i:
-// - If stride_order is empty, contiguity[i] corresponds to dimension i
-// - Otherwise, contiguity[i] corresponds to dimension
-// (rank - 1 - stride_order[i])
-std::vector<bool> getExpanded(
-    const std::vector<int64_t>& shape,
-    const std::vector<std::optional<bool>>& contiguity,
-    const std::vector<int64_t>& stride_order) {
-  size_t rank = shape.size();
-  std::vector<bool> is_expand(rank);
-  for (size_t index : arange(rank)) {
-    // since contiguity vector is given to the corresponding order in alloc
-    // domain, while is_expand is given to root domain, we need to map it
-    // correctly with `contig_index` and `index`.
-    //
-    // stride_order[i] indicates that:
-    //   `logical_domain[i]` maps to `alloc_domain[rank - 1 - stride_order_[i]]`
-    //
-    // Hence `index` on root domain would be corresponding to the contiguity
-    // index `contig_index = rank - 1 - stride_order[index]`
-    const size_t contig_index = stride_order.empty()
-        ? index
-        : rank - 1 - static_cast<size_t>(stride_order[index]);
-    const bool is_broadcast = !contiguity[contig_index].has_value();
-    const bool has_non_broadcast_size = (shape[index] != 1);
-    // A root dimension is expand dimension if:
-    //   The dimension is marked a broadcast; and
-    //   The dimension has an expanded extent.
-    is_expand[index] = is_broadcast && has_non_broadcast_size;
-  }
-  return is_expand;
-}
-
-// Normalizes and validates the stride order for a tensor.
-//
-// This function processes the stride order vector to ensure it's valid and
-// consistent:
-// 1. Handles negative indices by converting them to positive indices
-// 2. Validates that all indices are within the valid range
-// 3. Ensures there are no duplicate entries
-//
-// Parameters
-// ----------
-// stride_order : vector<int64_t>
-//     The stride order vector to normalize. Can be empty or contain indices
-//     from -rank to rank-1, where rank is the size of the vector.
-//
-// Notes
-// -----
-// - Negative indices are converted to positive by adding the rank
-// - All indices must be within the range [-rank, rank-1]
-// - No duplicate indices are allowed
-// - An empty stride_order vector is valid and will be left unchanged
-void normalizeStrideOrder(std::vector<int64_t>& stride_order) {
-  if (stride_order.empty()) {
-    return;
-  }
-  int64_t rank = (int64_t)stride_order.size();
-  std::unordered_set<int64_t> order_set;
-  for (auto& order : stride_order) {
-    order_set.insert(order);
-    if (order < 0) {
-      NVF_CHECK(
-          order >= -rank,
-          "defineTensor stride_order argument is out of range, expects >= ",
-          -rank,
-          ", but got: ",
-          order);
-      order += rank;
-    } else {
-      NVF_CHECK(
-          order < rank,
-          "defineTensor stride_order argument is out of range, expects < ",
-          rank,
-          ", but got: ",
-          order);
-    }
-  }
-  NVF_CHECK(
-      order_set.size() == stride_order.size(),
-      "defineTensor got duplicated stride_order entries: " +
-          toDelimitedString(stride_order));
-}
-
 // Creates a new TensorView with the specified properties.
 //
 // This function creates a tensor with the given shape, contiguity, data type,
